requirements.txt
Django==2.2.2
django-debug-toolbar==1.11
mysqlclient==1.4.2.post1
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
